fix(api): public nested data api correction
